Split FirstReservableTimeHelper and related classes into separate files
