+++ reservation_units/querysets.py
@@ -8,7 +8,7 @@ from elasticsearch_django.models import SearchResultsQuerySet
 
 from common.date_utils import local_datetime
 from common.db import ArrayUnnest, SubqueryArray
-from reservation_units.utils.first_reservable_time_helper import FirstReservableTimeHelper
+from reservation_units.utils.first_reservable_time_helper.first_reservable_time_helper import FirstReservableTimeHelper
 from resources.models import Resource
 from spaces.models import Space
