+++ reservation_units/utils/first_reservable_time_helper/utils.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, NamedTuple
+
+if TYPE_CHECKING:
+    from datetime import datetime
+
+
+class ReservableTimeOutput(NamedTuple):
+    is_closed: bool
+    first_reservable_time: datetime | None
